Beautifying code
