+++ App.js
@@ -37,13 +37,17 @@ import { Transition, FluidNavigator } from 'react-navigation-fluid-transitions'
 import { FlatGrid } from 'react-native-super-grid';
 import * as Progress from 'react-native-progress';
 import { PieChart } from 'react-native-svg-charts'
-import * as Data from './data/data'
+import * as Data from './src/data/countries'
 import Carousel from '@rhysforyou/react-native-carousel'
 import { translate, country } from "./src/localization"
 
+import Splash from "./src/screens/Splash"
+import Profile from "./src/screens/Profile"
+import ShoppingList from "./src/screens/ShoppingList"
+import SignUpModal from "./src/screens/SignUpModal"
+
 
 var { height, width } = Dimensions.get('window');
-console.log('Height', height)
 
 const calendarDayTextSize = height < 600 ? 14 : 15
 const calendarMonthTextSize = height < 600 ? 20 : 30
@@ -82,419 +86,6 @@ const applyMoneyMask = (quantity) => {
 	}
 
 	return str
-}
-
-class Splash extends Component {
-	constructor(props) {
-		super(props)
-		//161943 H&K
-		this.state = { code: "", storedCode: null, settings: {}, animError: new Animated.Value(0), spinner: new Animated.Value(0), errorMessage: "", busy: false }
-
-		var didBlurSubscription
-	}
-
-	_storeData = async (code, password) => {
-		console.log("STORE DATA")
-		try {
-			await AsyncStorage.setItem('@persistentItem:code', code);
-			await AsyncStorage.setItem('@persistentItem:password', password);
-			console.log("Saved!")
-		} catch (error) {
-			// Error saving data
-		}
-	};
-
-	_retrieveData = async (code, password) => {
-		console.log("RETRIEVE DATA")
-		try {
-			var pCode = await AsyncStorage.getItem('@persistentItem:code')
-			var pPassword = await AsyncStorage.getItem('@persistentItem:password')
-
-			code = pCode ? pCode : undefined || code;
-			password = pPassword ? pPassword : undefined || password;
-
-
-			if (code === undefined)
-				return Promise.reject("system_erro_code_needed")
-			if (password === undefined)
-				return Promise.reject("system_error_password_needed")
-
-
-			this.setState({ busy: true, code: code })
-
-			try {
-				let snapshot = await firebase.database().ref(`/users/${code}/`).once('value');
-
-				// console.log(snapshot.val().email)
-
-				if (snapshot.val() === null)
-					return Promise.reject(new Error("No data available using that code"))
-
-				const email = snapshot.val().email
-				console.log("password", password)
-
-				let user = await firebase.auth().signInWithEmailAndPassword(email, password)
-					.catch((reason) => { return Promise.reject(reason.code) })
-				console.log("LOGGED IN", user)
-
-				var moneyData = await firebase.database().ref(`/${code}/`)
-					.once('value')
-					.then((value) => {
-						return value.val() || { expenses: {}, shopping: [] }
-					})
-					.catch(errorMessage => {
-						console.log("Error", errorMessage)
-						return errorMessage
-					})
-
-
-				return moneyData
-
-				// return snapshot.val();
-			}
-			catch (errorCode) {
-				return Promise.reject(errorCode)
-			}
-
-		} catch (error) {
-			return Promise.reject("system_error_async_storage")
-		}
-	};
-
-	_retrieveSettings = async () => {
-		console.log("RETRIEVE SETTINGS")
-		try {
-			const value = await AsyncStorage.getItem('@persistentItem:currency');
-			console.log("Settings", value)
-
-			if (value === null)
-				return Promise.reject(new Error("No data"))
-
-			this.setState({ settings: value })
-			return value
-		} catch (error) {
-			// Error retrieving data
-			return error
-		}
-	};
-
-	componentDidMount() {
-
-		this.didBlurSubscription = this.props.navigation.addListener(
-			'didFocus',
-			payload => {
-			  this.setState({ busy: false, code: ""})
-			}
-		  );
-
-
-		this._retrieveSettings()
-			.then((value) => this._retrieveData()
-				.then(value => {
-					console.log("Retrieve data success", value)
-					this.props.navigation.navigate('Main', { moneyData: value, code: this.state.code, currency: this.state.settings })
-				})
-				.catch(error => {
-					this.animateErrorMessage(error)
-				}))
-			.catch((error) => this.props.navigation.navigate('Tutorial'))
-
-	}
-
-	componentWillUnmount(){
-		this.didBlurSubscription.remove()
-	}
-
-	onSubmit = (code, password) => {
-
-
-		this._retrieveSettings().then(
-			(value) => {
-				this._retrieveData(code, password)
-					.then((moneyData) =>
-						this._storeData(code, password)
-							.then(value => {
-								console.log("Retrieve data success SUBMIT1", moneyData)
-								this.props.navigation.navigate('Main', { moneyData: moneyData, code: this.state.code, currency: this.state.settings })
-							})
-					)
-					.catch((error) => {
-						this.animateErrorMessage(error)
-					})
-					.finally(() => this.setState({ busy: false }))
-			}
-		)
-	}
-
-	onCodeCreated = (code, password) => {
-		this.setState({ code: code.toString(), password: password })
-	}
-
-	animateErrorMessage = (message) => {
-
-		this.setState({ errorMessage: message })
-
-		const forwardsAnimation = Animated.spring(this.state.animError, {
-			toValue: 1,
-			duration: 200,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		const backwardsAnimation = Animated.spring(this.state.animError, {
-			toValue: 0,
-			duration: 300,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		forwardsAnimation.start()
-		setTimeout(() => { backwardsAnimation.start() }, 3000)
-
-	}
-
-	_renderForm = () => {
-
-		const { busy, code, password } = this.state
-
-		console.log("RENDER FORM?", busy)
-
-		return !busy
-			?
-			<View style={{ flexDirection: 'row', marginTop: 20 }}>
-				<View style={{ flex: 1 }} />
-				<View style={{ flex: 3 }}>
-					<TextInput onChangeText={(text) => this.setState({ code: text })} style={{ backgroundColor: 'white', borderRadius: 10, marginVertical: 10 }} keyboardType={'numeric'} value={this.state.code} placeholder={'Account Code'}></TextInput>
-					<TextInput onSubmitEditing={this.onSubmit.bind(this, code, password)} onChangeText={(text) => this.setState({ password: text })} style={{ backgroundColor: 'white', borderRadius: 10, marginVertical: 10 }} secureTextEntry={true} value={this.state.password} placeholder={'Password'}></TextInput>
-					<TouchableNativeFeedback onPress={this.onSubmit.bind(this, code, password)}
-						style={{ borderRadius: 20 }}>
-						<View style={{ alignItems: 'center', justifyContent: 'center', backgroundColor: BLU_LIGHT, paddingHorizontal: 20, borderRadius: 20, marginVertical: 10, paddingVertical: 10 }}>
-							<Text style={{ color: 'white' }}> {translate("splash_button")} </Text>
-						</View>
-					</TouchableNativeFeedback>
-					<TouchableNativeFeedback onPress={() => this.props.navigation.navigate('MyModal', { onCodeCreated: this.onCodeCreated })}>
-						<Text style={{ textDecorationLine: 'underline', color: BLU_LIGHT, textAlign: "center", marginVertical: 10 }}> Don't have an account? </Text>
-					</TouchableNativeFeedback>
-				</View>
-				<View style={{ flex: 1 }} />
-			</View>
-			: null
-	}
-
-
-	render() {
-		const { animError, errorMessage, busy } = this.state
-		return (
-			<View style={{ flex: 1, justifyContent: 'center', backgroundColor: BLU }}>
-				<View style={{ alignItems: 'center', justifyContent: 'center', }}>
-					<View style={{ backgroundColor: BLU, width: 160, height: 160, borderRadius: 80, borderColor: 'white', borderWidth: busy ? 0 : 4, alignItems: 'center', justifyContent: 'center' }}>
-						{busy ? <Progress.CircleSnail style={{ position: 'absolute' }} color={'white'} thickness={5} size={170} indeterminate={true} /> : null}
-						<Image resizeMode={'contain'} style={{ width: 150, height: 150, borderRadius: 150, position: "absolute" }} source={require('./img/logo.png')} />
-					</View>
-					<Text style={{ color: 'white', fontSize: 25, fontWeight: 'bold', fontFamily: 'Roboto', marginTop: 10 }}> Spendless</Text>
-				</View>
-				{this._renderForm()}
-				<Animated.View style={[{ backgroundColor: '#AA3C3B', borderRadius: 20, alignItems: 'center', justifyContent: 'center', flexDirection: 'row', marginHorizontal: width / 6, position: 'absolute', bottom: 0 }, {
-					transform: [{
-						translateY: animError.interpolate({
-							inputRange: [0, 1],
-							outputRange: [150, height * -.1]
-						})
-					}]
-				}]}>
-					<Text style={{ color: 'white', textAlign: 'center', padding: 10, flex: 1 }}>{translate(errorMessage)}</Text>
-				</Animated.View>
-			</View>
-		)
-	}
-}
-
-class ModalScreen extends React.Component {
-	constructor(props) {
-		super(props)
-		//161943 H&K
-		this.state = { email: "", password: "", animError: new Animated.Value(0), spinner: new Animated.Value(0), editable: true, done: false, code: null }
-
-	}
-
-	animateErrorMessage = (message) => {
-
-		this.setState({ errorMessage: message })
-
-		const forwardsAnimation = Animated.spring(this.state.animError, {
-			toValue: 1,
-			duration: 200,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		const backwardsAnimation = Animated.spring(this.state.animError, {
-			toValue: 0,
-			duration: 300,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		forwardsAnimation.start()
-		setTimeout(() => { backwardsAnimation.start() }, 3000)
-
-	}
-
-	onSubmit = (email, password) => {
-
-
-		const forwardsAnimation = Animated.spring(this.state.spinner, {
-			toValue: 1,
-			duration: 500,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		const backwardsAnimation = Animated.spring(this.state.spinner, {
-			toValue: 0,
-			duration: 1000,
-			friction: 8,
-			tension: 50,
-			useNativeDriver: true
-		})
-
-		this.setState({ editable: false })
-
-		if (password !== this.state.repeatPassword) {
-			this.animateErrorMessage("system_error_passwords_match")
-			this.setState({ editable: true })
-			return
-		}
-
-
-		forwardsAnimation.start()
-
-		setTimeout(() => {
-			firebase.auth()
-				.createUserWithEmailAndPassword(email, password)
-				.then((value) => {
-
-
-
-					var code = Math.floor(100000 + Math.random() * 900000)
-					var userData = { email: email, dateCreated: new Date().getTime(), country: country }
-
-					this.pushData(`/users/${code}/`, userData, () => {
-						backwardsAnimation.start()
-						LayoutAnimation.configureNext({
-							duration: 700,
-							create: {
-								type: LayoutAnimation.Types.spring,
-								property: LayoutAnimation.Properties.scaleXY,
-								springDamping: 1,
-								duration: 600
-							},
-							update: {
-								type: LayoutAnimation.Types.spring,
-								property: LayoutAnimation.Properties.scaleXY,
-								springDamping: 1,
-								duration: 600
-							},
-							delete: {
-								type: LayoutAnimation.Types.easeOut,
-								property: LayoutAnimation.Properties.opacity,
-								springDamping: 1,
-								duration: 400
-							}
-						})
-
-						this.setState({ code: code })
-					})
-				})
-				.catch((reason) => {
-					backwardsAnimation.start()
-					console.log("Error", reason)
-					setTimeout(() => { this.animateErrorMessage(reason.code) }, 200)
-					// this.animateErrorMessage(reason.message)
-				})
-				.finally(() => this.setState({ editable: true }))
-		}, 1500)
-
-	}
-
-	pushData = async (ref, value, callback) => {
-		console.log(ref)
-		let snapshot = await firebase.database().ref(ref)
-			.set(value)
-			.then((value) => callback())
-
-		console.log(snapshot.val())
-	}
-
-	render() {
-		const { animError, errorMessage, spinner, editable, code, password } = this.state
-		var accountCreated = code !== null;
-		return (
-			<View style={{ flex: 1, backgroundColor: BLU, justifyContent: "center", alignItems: "center" }}>
-				<Text style={{ color: "white", textAlign: 'center', fontSize: 30, paddingHorizontal: 20 }}>{translate( accountCreated ? "account_title_done" : "account_title")}</Text>
-				<View style={{ flexDirection: 'row', marginTop: 20 }}>
-					<View style={{ flex: 1 }} />
-					{accountCreated ?
-						<View style={{ flex: 3 }}>
-							<View style={{ justifyContent: 'center', alignItems: "center", marginVertical: 20 }}>
-								<View style={{ width: 60, height: 60, borderRadius: 30, alignItems: "center", justifyContent: 'center', backgroundColor: '#afc474' }}>
-									<Icon color={'white'} size={60} name="check-circle" />
-								</View>
-							</View>
-							<Text style={{ color: "white", textAlign: 'center', fontSize: 18 }}>{translate("account_code")}</Text>
-							<Text style={{ color: "white", fontSize: 30, marginVertical: 10, textAlign: 'center' }}>{code.toString()}</Text>
-							<Text style={{ color: "white", paddingHorizontal: 5, textAlign: 'center', fontSize: 18 }}>{translate("account_share")}</Text>
-							<TouchableNativeFeedback onPress={(event) => { this.props.navigation.state.params.onCodeCreated(code, password); this.props.navigation.goBack() }}
-								style={{ borderRadius: 20 }}>
-								<View style={{ alignItems: 'center', justifyContent: 'center', backgroundColor: BLU_LIGHT, paddingHorizontal: 20, borderRadius: 20, marginVertical: 20, paddingVertical: 10 }}>
-									<Text style={{ color: 'white' }}> {translate("account_button")} </Text>
-								</View>
-							</TouchableNativeFeedback>
-						</View> :
-						<View style={{ flex: 3 }}>
-							<Text style={{ color: "white" }}>{translate("account_email")}</Text>
-							<TextInput editable={editable} onChangeText={(text) => this.setState({ email: text })} style={{ backgroundColor: 'white', borderRadius: 10, marginVertical: 5 }} keyboardType={translate("account_email_hint")} value={this.state.email} placeholder={'Email'}></TextInput>
-							<Text style={{ color: "white", marginTop: 10 }}>{translate("account_password")}</Text>
-							<TextInput editable={editable} onChangeText={(text) => this.setState({ password: text })} style={{ backgroundColor: 'white', borderRadius: 10, marginVertical: 5 }} secureTextEntry={true} value={this.state.password} placeholder={translate("account_password_hint")}></TextInput>
-							<Text style={{ color: "white", marginTop: 10 }}> { translate("account_repeat_password")} </Text>
-							<TextInput editable={editable} onChangeText={(text) => this.setState({ repeatPassword: text })} style={{ backgroundColor: 'white', borderRadius: 10, marginVertical: 5 }} secureTextEntry={true} value={this.state.repeatPassword} placeholder={translate("account_repeat_password_hint")}></TextInput>
-							<TouchableNativeFeedback onPress={this.onSubmit.bind(this, this.state.email, this.state.password)}
-								style={{ borderRadius: 20 }}>
-								<View style={{ alignItems: 'center', justifyContent: 'center', backgroundColor: BLU_LIGHT, paddingHorizontal: 20, borderRadius: 20, marginVertical: 20, paddingVertical: 10 }}>
-									<Text style={{ color: 'white' }}> Sign Up </Text>
-								</View>
-							</TouchableNativeFeedback>
-						</View>}
-					<View style={{ flex: 1 }} />
-				</View>
-				<Animated.View style={{
-					position: "absolute", bottom: 0, transform: [{
-						translateY: spinner.interpolate({
-							inputRange: [0, 1],
-							outputRange: [height + 150, height * -.1]
-						})
-					}]
-				}}>
-					<Progress.Circle thickness={15} size={50} indeterminate={true} />
-				</Animated.View>
-				<Animated.View style={[{ backgroundColor: '#AA3C3B', borderRadius: 20, alignItems: 'center', justifyContent: 'center', flexDirection: 'row', marginHorizontal: width / 6, position: 'absolute', bottom: 0 }, {
-					transform: [{
-						translateY: animError.interpolate({
-							inputRange: [0, 1],
-							outputRange: [150, height * -.1]
-						})
-					}]
-				}]}>
-					<Text style={{ color: 'white', textAlign: 'center', padding: 10, flex: 1 }}>{translate(errorMessage)}</Text>
-				</Animated.View>
-			</View>
-		);
-	}
 }
 
 class Currency extends Component {
@@ -1274,48 +865,6 @@ class History extends Component {
 	}
 }
 
-class Profile extends Component {
-	constructor(props) {
-		super(props)
-		this.state = { refresh: false }
-
-		this.code = this.props.navigation.getParam('code', null)
-		this.user = firebase.auth().currentUser
-
-		console.log(this.code)
-	}
-
-	componentDidMount() {
-
-
-	}
-
-	render() {
-		return (
-			<View style={{ flex: 1, flexDirection: 'column', backgroundColor: BLU, paddingHorizontal: 20 }}>
-				<Text style={{ color: 'white', fontSize: 35, marginVertical: 20 }}>Profile</Text>
-				<Text style={{ color: 'white', fontSize: 20 }}>Account Email</Text>
-				<Text style={{ color: 'white', fontSize: 17 }}>{this.user.email}</Text>
-				<Text style={{ color: 'white', fontSize: 20, marginTop: 10 }}>Account Code</Text>
-				<Text style={{ color: 'white', fontSize: 17 }}>{this.code}</Text>
-				<View style={{ flex: 1 }} />
-				<TouchableNativeFeedback onPress={(event) => {
-					firebase.auth().signOut()
-						.then((value) => { 
-							AsyncStorage.multiRemove(["@persistentItem:password","@persistentItem:code"])
-							.then((value) => { this.props.navigation.dispatch(StackActions.popToTop()) }) 
-						})
-				}}
-					style={{ borderRadius: 20 }}>
-					<View style={{ alignItems: 'center', justifyContent: 'center', backgroundColor: "#AA3C3B", paddingHorizontal: 20, borderRadius: 20, marginVertical: 20, paddingVertical: 10 }}>
-						<Text style={{ color: 'white' }}> Log out </Text>
-					</View>
-				</TouchableNativeFeedback>
-			</View>
-		)
-	}
-}
-
 class ShoppingList extends Component {
 	constructor(props) {
 		super(props)
@@ -1565,8 +1114,8 @@ const RootStack = createStackNavigator(
 		AppNavigator: {
 			screen: AppNavigator,
 		},
-		MyModal: {
-			screen: ModalScreen,
+		SignUpModal: {
+			screen: SignUpModal,
 		},
 	},
 	{
