+++ src/data/countries.js
@@ -4,247 +4,247 @@ export const countries = [
         name: 'Australia',
         currency: 'AUD',
         currency_name: 'Dollar',
-        image: require('../img/au.png')
+        image: require('../../img/au.png')
     },
     {
         id: 'br',
         name: 'Brazil',
         currency: 'BRL',
         currency_name: 'Real',
-        image: require('../img/br.png')
+        image: require('../../img/br.png')
     },
     {
         id: 'ca',
         name: 'Canada',
         currency: 'USD',
         currency_name: 'Dollar',
-        image: require('../img/ca.png')
+        image: require('../../img/ca.png')
     },
     {
         id: 'cl',
         name: 'Chile',
         currency: 'CLP',
         currency_name: 'Peso',
-        image: require('../img/cl.png')
+        image: require('../../img/cl.png')
     },
     {
         id: 'cn',
         name: 'China',
         currency: 'CNY',
         currency_name: 'Renminbi',
-        image: require('../img/cn.png')
+        image: require('../../img/cn.png')
     },
     {
         id: 'co',
         name: 'Colombia',
         currency: 'COL',
         currency_name: 'Peso',
-        image: require('../img/co.png')
+        image: require('../../img/co.png')
     },
     {
         id: 'cz',
         name: 'Czechia',
         currency: 'CZK',
         currency_name: 'Koruna',
-        image: require('../img/cz.png')
+        image: require('../../img/cz.png')
     },
     {
         id: 'dk',
         name: 'Denmark',
         currency: 'DKK',
         currency_name: 'Krone',
-        image: require('../img/dk.png')
+        image: require('../../img/dk.png')
     },
     {
         id: 'eu',
         name: 'European Union',
         currency: 'EUR',
         currency_name: 'Euro',
-        image: require('../img/eu.png')
+        image: require('../../img/eu.png')
     },
     {
         id: 'gb',
         name: 'Great Britain',
         currency: 'GBP',
         currency_name: 'Pound Sterling',
-        image: require('../img/gb.png')
+        image: require('../../img/gb.png')
     },
     {
         id: 'hk',
         name: 'Hong Kong',
         currency: 'HKD',
         currency_name: 'Dollar',
-        image: require('../img/hk.png')
+        image: require('../../img/hk.png')
     },
     {
         id: 'hu',
         name: 'Hungary',
         currency: 'HUF',
         currency_name: 'Forint',
-        image: require('../img/hu.png')
+        image: require('../../img/hu.png')
     },
     {
         id: 'is',
         name: 'Iceland',
         currency: 'ISK',
         currency_name: 'Krona',
-        image: require('../img/is.png')
+        image: require('../../img/is.png')
     },
     {
         id: 'in',
         name: 'India',
         currency: 'INR',
         currency_name: 'Rupee',
-        image: require('../img/in.png')
+        image: require('../../img/in.png')
     },
     {
         id: 'id',
         name: 'Indonesia',
         currency: 'IDR',
         currency_name: 'Rupiah',
-        image: require('../img/id.png')
+        image: require('../../img/id.png')
     }, {
         id: 'il',
         name: 'Israel',
         currency: 'ILS',
         currency_name: 'New Shekel',
-        image: require('../img/il.png')
+        image: require('../../img/il.png')
     },
     {
         id: 'jp',
         name: 'Japan',
         currency: 'JPY',
         currency_name: 'Yen',
-        image: require('../img/jp.png')
+        image: require('../../img/jp.png')
     },
     {
         id: 'my',
         name: 'Malaysia',
         currency: 'MYR',
         currency_name: 'Ringgit',
-        image: require('../img/my.png')
+        image: require('../../img/my.png')
     },
     {
         id: 'mx',
         name: 'Mexico',
         currency: 'MXN',
         currency_name: 'Peso',
-        image: require('../img/mx.png')
+        image: require('../../img/mx.png')
     },
     {
         id: 'nz',
         name: 'New Zeland',
         currency: 'NZD',
         currency_name: 'Dollar',
-        image: require('../img/nz.png')
+        image: require('../../img/nz.png')
     },
     {
         id: 'no',
         name: 'Norway',
         currency: 'NOK',
         currency_name: 'Krone',
-        image: require('../img/no.png')
+        image: require('../../img/no.png')
     },
     {
         id: 'ph',
         name: 'The Phillipines',
         currency: 'PHP',
         currency_name: 'Peso',
-        image: require('../img/ph.png')
+        image: require('../../img/ph.png')
     },
     {
         id: 'pl',
         name: 'Poland',
         currency: 'PLN',
         currency_name: 'Zloty',
-        image: require('../img/pl.png')
+        image: require('../../img/pl.png')
     },
     {
         id: 'ro',
         name: 'Romania',
         currency: 'RON',
         currency_name: 'Leu',
-        image: require('../img/ro.png')
+        image: require('../../img/ro.png')
     }, {
         id: 'ru',
         name: 'Russia',
         currency: 'RUB',
         currency_name: 'Ruble',
-        image: require('../img/ru.png')
+        image: require('../../img/ru.png')
     },
     {
         id: 'sa',
         name: 'Saudi Arabia',
         currency: 'SAR',
         currency_name: 'Riyal',
-        image: require('../img/sa.png')
+        image: require('../../img/sa.png')
     },
     {
         id: 'sg',
         name: 'Singapore',
         currency: 'SGD',
         currency_name: 'Dollar',
-        image: require('../img/sg.png')
+        image: require('../../img/sg.png')
     },
     {
         id: 'za',
         name: 'South Africa',
         currency: 'ZAR',
         currency_name: 'Rand',
-        image: require('../img/za.png')
+        image: require('../../img/za.png')
     }, {
         id: 'kr',
         name: 'South Korea',
         currency: 'KRW',
         currency_name: 'Won',
-        image: require('../img/kr.png')
+        image: require('../../img/kr.png')
     },
     {
         id: 'se',
         name: 'Sweden',
         currency: 'SEK',
         currency_name: 'Krona',
-        image: require('../img/se.png')
+        image: require('../../img/se.png')
     },
     {
         id: 'ch',
         name: 'Switzerland',
         currency: 'CHF',
         currency_name: 'Franc',
-        image: require('../img/ch.png')
+        image: require('../../img/ch.png')
     },
     {
         id: 'tw',
         name: 'Taiwan',
         currency: 'TWD',
         currency_name: 'Dollar',
-        image: require('../img/tw.png')
+        image: require('../../img/tw.png')
     }, {
         id: 'th',
         name: 'Thailand',
         currency: 'THB',
         currency_name: 'Baht',
-        image: require('../img/th.png')
+        image: require('../../img/th.png')
     },
     {
         id: 'tr',
         name: 'Turkey',
         currency: 'TRY',
         currency_name: 'Lira',
-        image: require('../img/tr.png')
+        image: require('../../img/tr.png')
     },
     {
         id: 'ae',
         name: 'United Arab Emirates',
         currency: 'AED',
         currency_name: 'Dirham',
-        image: require('../img/ae.png')
+        image: require('../../img/ae.png')
     },
     {
         id: 'us',
         name: 'United States of America',
         currency: 'USD',
         currency_name: 'Dollar',
-        image: require('../img/us.png')
+        image: require('../../img/us.png')
     }
 ]
