+++ src/localization/en.json
@@ -14,6 +14,7 @@
     "currency_button": "Select",
     "currency_message": "Select your preferred currency",
     "splash_button": "Submit",
+    "splash_dont_have": "Don't have an account?",
     "main_category_vegetables": "Vegetables",
     "main_category_fruits": "Fruits",
     "main_category_dairy": "Dairy",
@@ -32,13 +33,18 @@
     "main_add_expense_hint": "Amount",
     "history_amount": "Amount",
     "history_time": "Time",
+    "profile_title": "Profile",
+    "profile_email": "Account Email",
+    "profile_account": "Account Code",
+    "profile_button_log_out": "Log Out",
     "shopping_list_title": "Shopping List",
     "shopping_list_hint": "New Item",
     "account_title": "Create an Account",
     "account_title_done": "Account Succesfully Created!",
     "account_code": "Your account code is:",
     "account_share": "You can share this code with others to access the same account",
-    "account_button": "LOG IN",
+    "account_button_log_in": "LOG IN",
+    "account_button_sign_up": "SIGN UP",
     "account_email": "Email",
     "account_password": "Password",
     "account_repeat_password": "Repeat Password",
