+++ src/localization/es.json
@@ -1,5 +1,7 @@
 {
     "tutorial_header_title": "Bienvenido a Spendless!",
+    "tutorial_card_0_title": "Crea una Cuenta",
+    "tutorial_card_0_description": "¡Comparte tu Código con otros para poder utilizar Spendless al máximo! Podrás compartir tus gastos y lista de compras",
     "tutorial_card_1_title": "Una Interfaz Sencilla",
     "tutorial_card_1_description": "Selecciona cualquier día y comienza a registrar tus gastos",
     "tutorial_card_2_title": "Registra tus gastos día con día",
@@ -12,6 +14,7 @@
     "currency_button": "Seleccionar",
     "currency_message": "Selecciona tu moneda preferida",
     "splash_button": "Acceder",
+    "splash_dont_have": "¿No tienes una cuenta?",
     "main_category_vegetables": "Vegetales",
     "main_category_fruits": "Frutas",
     "main_category_dairy": "Lácteos",
@@ -30,13 +33,18 @@
     "main_add_expense_hint": "Cantidad",
     "history_amount": "Cantidad",
     "history_time": "Tiempo",
+    "profile_title": "Perfil",
+    "profile_email": "Correo",
+    "profile_account": "Código de Cuenta",
+    "profile_button_log_out": "Cerrar sesión",
     "shopping_list_title": "Lista de Compras",
     "shopping_list_hint": "Nuevo Artículo",
     "account_title": "Crea una Cuenta",
     "account_title_done": "¡Cuenta Creada Exitosamente!",
     "account_code": "Tu código es:",
     "account_share": "Puedes compartir este código con otros para acceder a la misma cuenta",
-    "account_button": "INICIAR SESIÓN",
+    "account_button_log_in": "INICIAR SESIÓN",
+    "account_button_sign_up": "CREAR CUENTA",
     "account_email": "Correo Electrónico",
     "account_password": "Contraseña",
     "account_repeat_password": "Repetir Contraseña",
